+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ---
 
-## 7.0.2 (TBD)
+## [7.0.2](https://github.com/ngageoint/geopackage-ios/releases/tag/7.0.2) (12-16-2021)
 
 * GPKGUrlTileGenerator HTTP Method and Header field values support
 * TileCreator improved tile sizes when tile width and/or height are not specified